+++ js/formValidation.js
@@ -4,51 +4,55 @@
  * @requires module:util
  */
 import { checkValue, checkTimeInput, parseAge } from './util.js';
+import * as LOG from './logger.js';
 
 /**
  * Set up event listeners for form validation
  * @param   {FormValidationConfig[]} config  configuration for setup
+ * @param   {Boolean}                logger  Set to true to enable logger
  * @returns {String} space-delimited list of all selectors used
  */
 export default function setup(config) {
-  let fields = "";
+  LOG.beginFunction('[formValidation] setup', arguments);
+  let fields = '';
   for ( const item of config ) {
     fields += `${item.selector}, `;
     // Remove 'invalid' class on focus
-    $(item.selector).on("focus", e => {
-      $(e.target).removeClass("invalid");
+    $(item.selector).on('focus', e => {
+      $(e.target).removeClass('invalid');
     });
     // Handle items with an itemType
     if ( item.inputType ) {
-      if ( item.inputType === "age" ) {
-        $(item.selector).on("focusout", e => {
-          if ( $(e.target).val() !== "" ) {
+      if ( item.inputType === 'age' ) {
+        $(item.selector).on('focusout', e => {
+          if ( $(e.target).val() !== '' ) {
             validateAge(e.target, item);
           }
         });
-      } else if ( item.inputType === "time" ) {
-        $(item.selector).on("focusout", e => {
-          if ( $(e.target).val() !== "" ) {
+      } else if ( item.inputType === 'time' ) {
+        $(item.selector).on('focusout', e => {
+          if ( $(e.target).val() !== '' ) {
             validateTime(e.target, item);
           }
         });
       }
     } else if ( item.max ) {
       // Handle items with a range
-      $( item.selector ).on("focusout", e => {
-        if ( $( e.target ).val() !== "" ) {
+      $( item.selector ).on('focusout', e => {
+        if ( $( e.target ).val() !== '' ) {
           validateRange(e.target, item);
         }
       });
     } else if ( item.match ) {
       // Handle items with a RegExp match
-      $(item.selector).on("focusout", e => {
-        if ( $(e.target).val() !== "" ) {
+      $(item.selector).on('focusout', e => {
+        if ( $(e.target).val() !== '' ) {
           validateMatch(e.target, item);
         }
       });
     }
   }
+  LOG.endResult(fields.slice(0, -2));
   // Remove trailing comma-space
   return fields.slice(0, -2);
 }
@@ -74,14 +78,18 @@ export default function setup(config) {
  * @returns {HTMLElement}                 The original DOM element, for chaining
  */
 function validateAge(el, item) {
+  LOG.beginFunction('[formValidation] validateAge', arguments);
   const yearsOld = parseAge( $(el).val() );
-
+  LOG.blueText(`${$(el).val()} = ${yearsOld} years`);
   const validatedAge = checkValue(yearsOld, item.min, item.max);
   if ( validatedAge === 0 ) {
-    $(el).addClass("invalid");
+    LOG.redText(`Invalid input: ${$(el).val()}`);
+    $(el).addClass('invalid');
   } else {
-    $(el).removeClass("invalid");
+    LOG.greenText(`Valid input: ${$(el).val()}`);
+    $(el).removeClass('invalid');
   }
+  LOG.groupEnd();
   return el;
 }
 /**
@@ -95,15 +103,19 @@ function validateAge(el, item) {
  */
 /* eslint-disable-next-line no-unused-vars */
 function validateTime(el, item) {
+  LOG.beginFunction('[formValidation] validateTime', arguments);
   const x = $(el).val();
   const corrected = checkTimeInput(x);
-  if ( corrected === "" ) {
-    $(el).addClass("invalid");
+  if ( corrected === '' ) {
+    LOG.redText(`Invalid time input: ${x}`);
+    $(el).addClass('invalid');
   } else {
+    if ( corrected !== x ) LOG.blueText(`${x} corrected to ${corrected}`);
     $(el).val(corrected);
-    $(el).removeClass('invalid');
+    LOG.greenText(`Valid time input: ${corrected}`);
     $(el).removeClass('invalid');
   }
+  LOG.groupEnd();
   return el;
 }
 /**
@@ -115,11 +127,15 @@ function validateTime(el, item) {
  * @returns {HTMLElement}                The original DOM element, for chaining
  */
 function validateRange(el, item) {
+  LOG.beginFunction('[formValidation] validateRange', arguments);
   if ( checkValue(+$(el).val(), item.min, item.max) === 0 ) {
-    $(el).addClass("invalid");
+    LOG.redText(`Invalid input: ${$(el).val()}`);
+    $(el).addClass('invalid');
   } else {
+    LOG.greenText(`Valid input: ${$(el).val()}`);
     $(el).removeClass('invalid');
   }
+  LOG.groupEnd();
   return el;
 }
 /**
@@ -130,10 +146,14 @@ function validateRange(el, item) {
  * @returns {HTMLElement}                The original DOM element, for chaining
   */
 function validateMatch(el, item) {
+  LOG.beginFunction('[formValidation] validateMatch', arguments);
   if ( !item.match.test($(el).val())) {
-    $(el).addClass("invalid");
+    LOG.redText(`Invalid input: ${$(el).val()}`);
+    $(el).addClass('invalid');
   } else {
+    LOG.greenText(`Valid input: ${$(el).val()}`);
     $(el).removeClass('invalid');
   }
+  LOG.groupEnd();
   return el;
 }